backend/requirements.txt
# FastAPI và web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic cho validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Big Data clients
hdfs==2.7.3
kafka-python==2.0.2
pymongo==4.6.1

# Image processing
Pillow==10.2.0

# Utilities
python-dotenv==1.0.0

# Optional: Authentication (nếu cần JWT sau này)
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4
